+++ lib/src/services/validation_service.dart
@@ -1,10 +1,13 @@
 #!/usr/bin/env dart
 
 import 'package:adapters_flutter/src/managers/api_manager_.dart';
+import 'package:adapters_flutter/src/managers/log_manager.dart';
 import 'package:adapters_flutter/src/models/config_model.dart';
 import 'package:adapters_flutter/src/models/exception_model.dart';
 import 'package:meta/meta.dart';
 import 'package:uuid/uuid.dart';
+
+final _logger = getLogger();
 
 @internal
 void validateConfig(final ConfigModel? config) {
@@ -14,70 +17,66 @@ void validateConfig(final ConfigModel? config) {
   }
 
   if (config == null) {
-    throw TmsConfigException('Config is invalid: "$config".');
+    _logAndThrow('Config is invalid: "$config".');
   }
 
-  if (config.adapterMode == null ||
+  if (config!.adapterMode == null ||
       config.adapterMode! < 0 ||
       config.adapterMode! > 2) {
-    throw TmsConfigException(
-        'Adapter mode is invalid: "${config.adapterMode}".');
+    _logAndThrow('Adapter mode is invalid: "${config.adapterMode}".');
   }
 
   if (config.automaticCreationTestCases == null) {
-    throw TmsConfigException(
+    _logAndThrow(
         'Automatic creation test cases flag is invalid: "${config.automaticCreationTestCases}".');
   }
 
   if (config.automaticUpdationLinksToTestCases == null) {
-    throw TmsConfigException(
+    _logAndThrow(
         'Automatic updation links to test cases flag is invalid: "${config.automaticUpdationLinksToTestCases}".');
   }
 
   if (config.certValidation == null) {
-    throw TmsConfigException(
+    _logAndThrow(
         'Certificate validation flag is invalid: "${config.certValidation}".');
   }
 
   if (config.configurationId == null ||
       !Uuid.isValidUUID(fromString: config.configurationId!)) {
-    throw TmsConfigException(
-        'Configuration id is invalid: "${config.configurationId}".');
+    _logAndThrow('Configuration id is invalid: "${config.configurationId}".');
   }
 
   if (config.isDebug == null) {
-    throw TmsConfigException('Is debug flag is invalid: "${config.isDebug}".');
+    _logAndThrow('Is debug flag is invalid: "${config.isDebug}".');
   }
 
   if (config.privateToken == null || config.privateToken!.isEmpty) {
-    throw TmsConfigException(
-        'Private token is invalid: "${config.privateToken}".');
+    _logAndThrow('Private token is invalid: "${config.privateToken}".');
   }
 
   if (config.projectId == null ||
       !Uuid.isValidUUID(fromString: config.projectId!)) {
-    throw TmsConfigException('Project id is invalid: "${config.projectId}".');
+    _logAndThrow('Project id is invalid: "${config.projectId}".');
   }
 
   if (config.testIt == null) {
-    throw TmsConfigException('Test IT flag is invalid: "${config.testIt}".');
+    _logAndThrow('Test IT flag is invalid: "${config.testIt}".');
   }
 
   if (config.adapterMode == 0 || config.adapterMode == 1) {
     if (config.testRunId == null ||
         !Uuid.isValidUUID(fromString: config.testRunId!)) {
-      throw TmsConfigException(
-          'Test run id is invalid: "${config.testRunId}".');
+      _logAndThrow('Test run id is invalid: "${config.testRunId}".');
     }
   } else if (config.adapterMode == 2) {
     if (config.testRunId != null && config.testRunId!.isNotEmpty) {
-      throw TmsConfigException(
+      _logAndThrow(
           'TestRunId should be absent in adapter mode 2, but was "${config.testRunId}".');
     }
   }
 
   if (config.url == null || !(Uri.tryParse(config.url!)?.isAbsolute ?? false)) {
-    throw TmsConfigException('Url is invalid: "${config.url}".');
+    _logAndThrow('Url is invalid: "${config.url}".');
   }
 }
 
@@ -107,4 +106,11 @@ Future<void> validateWorkItemsIdsAsync(
 
   throw TmsArgumentException(
       'WorkItem with id "$notFoundWorkItemId" not found.');
+}
+
+void _logAndThrow(final String message) {
+  final exception = TmsConfigException(message);
+  _logger.e(exception);
+
+  throw exception;
 }