+++ test/test_manager_test.dart
@@ -4,9 +4,14 @@ import 'package:adapters_flutter/adapters_flutter.dart';
 
 void main() {
   group('test manager', () {
-    test('run tms test', () {
+    test('tms test', () {
       expect(
           () => tmsTest('', () {}), throwsA(const TypeMatcher<StateError>()));
+    });
+
+    test('tms test widgets', () {
+      expect(() => tmsTestWidgets('', (tester) async {}),
+          throwsA(const TypeMatcher<StateError>()));
     });
   });
 }