+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.  
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [1.8.0] - 2024-07-09
+
+### Added
+
+- tmsTestWidgets function.
+- validation error log.
+
+### Changed
+
+- examples.
+- readme.
+- unit test.
+
 ## [1.7.1] - 2024-07-08
 
 ### Added